v.3.2.1

+++ lin_guider/src/video_dev/asi_core.cpp
@@ -412,7 +412,7 @@ bool asi_core::set_camera_exposure(long exp_time) {
 
 bool asi_core::set_camera_gain(unsigned int gain) {
 	if (!m_has_gain) return false;
-	if((gain < m_gain_caps.MinValue) || (gain > m_gain_caps.MaxValue)) {
+	if(((int)gain < m_gain_caps.MinValue) || ((int)gain > m_gain_caps.MaxValue)) {
 		log_e("Gain %d not supported", gain);
 		return false;
 	}
